:white_check_mark: 광고추가 ok
ok
